pyproject.toml
[tool.poetry]
name = "mcadmin"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
psycopg = {extras = ["binary"], version = "^3.2.1"}
environs = {extras = ["django"], version = "^11.0.0"}
whitenoise = "^6.7.0"
rcon = "^2.4.6"
httpx = "^0.27.0"
django-compressor = "^4.5"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
